rename S3ImageUtil.kt to S3ImageUploader.kt
