+++ src/main/kotlin/kr/mashup/bangwidae/asked/external/aws/S3ImageUploader.kt
@@ -1,4 +1,4 @@
-package kr.mashup.bangwidae.asked.utils
+package kr.mashup.bangwidae.asked.external.aws
 
 import com.amazonaws.services.s3.AmazonS3
 import com.amazonaws.services.s3.model.CannedAccessControlList
@@ -10,7 +10,7 @@ import org.springframework.web.multipart.MultipartFile
 import java.util.*
 
 @Component
-class S3ImageUtil(
+class S3ImageUploader(
     private val amazonS3: AmazonS3,
     @Value("\${cloud.aws.s3.bucket}") private val bucket: String
 ) {
