+++ src/main/kotlin/kr/mashup/bangwidae/asked/service/UserService.kt
@@ -8,8 +8,8 @@ import kr.mashup.bangwidae.asked.exception.DoriDoriException
 import kr.mashup.bangwidae.asked.exception.DoriDoriExceptionType
 import kr.mashup.bangwidae.asked.model.User
 import kr.mashup.bangwidae.asked.repository.UserRepository
-import kr.mashup.bangwidae.asked.utils.S3ImageUtil
-import kr.mashup.bangwidae.asked.utils.UploadDirName
+import kr.mashup.bangwidae.asked.external.aws.S3ImageUploader
+import kr.mashup.bangwidae.asked.external.aws.UploadDirName
 import org.bson.types.ObjectId
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
@@ -21,7 +21,7 @@ class UserService(
     private val userRepository: UserRepository,
     private val certMailService: CertMailService,
     private val passwordService: PasswordService,
-    private val s3ImageUtil: S3ImageUtil
+    private val s3ImageUploader: S3ImageUploader
 ) {
 
     fun joinUser(joinUserRequest: JoinUserRequest): JoinUserResponse {
@@ -66,7 +66,7 @@ class UserService(
     }
 
     fun updateProfileImage(user: User, image: MultipartFile): String {
-        val uploadedUrl = s3ImageUtil.upload(image, UploadDirName.PROFILE)
+        val uploadedUrl = s3ImageUploader.upload(image, UploadDirName.PROFILE)
         userRepository.save(user.updateProfileImageUrl(uploadedUrl))
         return uploadedUrl
     }
